services/patch-drafter/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
google-generativeai>=0.7.0
redis>=5.0.3
sqlalchemy[asyncio]>=2.0.28
asyncpg>=0.29.0
pydantic>=2.0.0
